pyproject.toml
[tool.poetry]
name = "airflow-example"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
acryl-datahub = {extras = ["airflow"], version = "^0.8.15"}
psycopg2 = "^2.9.1"
apache-airflow = {extras = ["celery"], version = "^2.1.4"}
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
